requirements.txt
openai
requests
beautifulsoup4
python-docx
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
